+++ src/Data/Aeson/Lens.hs
@@ -73,6 +73,11 @@ import qualified Data.Text.Lazy.Encoding as LazyText
 import Data.Vector (Vector)
 import Prelude hiding (null)
 
+#if MIN_VERSION_aeson(2,0,0)
+import qualified Data.Aeson.Key    as Key
+import qualified Data.Aeson.KeyMap as KM
+#endif
+
 -- $setup
 -- >>> import Control.Lens
 -- >>> import Data.Aeson
@@ -296,7 +301,15 @@ class AsPrimitive t => AsValue t where
   -- >>> _Object._Wrapped # [("key" :: Text, _String # "value")] :: String
   -- "{\"key\":\"value\"}"
   _Object :: Prism' t (HashMap Text Value)
-  _Object = _Value.prism Object (\v -> case v of Object o -> Right o; _ -> Left v)
+  _Object = _Value.prism (Object . fwd) (\v -> case v of Object o -> Right (bwd o); _ -> Left v)
+    where
+#if MIN_VERSION_aeson(2,0,0)
+      fwd = KM.fromHashMapText
+      bwd = KM.toHashMapText
+#else
+      fwd = id
+      bwd = id
+#endif
   {-# INLINE _Object #-}
 
   -- |
@@ -472,7 +485,14 @@ type instance Index Value = Text
 
 type instance IxValue Value = Value
 instance Ixed Value where
-  ix i f (Object o) = Object <$> ix i f o
+  ix i f (Object o) = Object <$> ix (toKey i) f o
+    where
+#if MIN_VERSION_aeson(2,0,0)
+      toKey :: Text -> Key.Key
+      toKey = Key.fromText
+#else
+      toKey = id
+#endif
   ix _ _ v          = pure v
   {-# INLINE ix #-}
 
@@ -481,6 +501,21 @@ instance Plated Value where
   plate f (Array a) = Array <$> traverse f a
   plate _ xs = pure xs
   {-# INLINE plate #-}
+
+#if MIN_VERSION_aeson(2,0,0)
+type instance Index (KM.KeyMap v) = Key.Key
+type instance IxValue (KM.KeyMap v) = v
+
+instance Ixed (KM.KeyMap v)
+
+instance At (KM.KeyMap v) where
+  at k f = KM.alterF f k
+  {-# INLINE at #-}
+
+instance Each (KM.KeyMap a) (KM.KeyMap b) a b where
+  each = traversed
+  {-# INLINE each #-}
+#endif
 
 ------------------------------------------------------------------------------
 -- Pattern Synonyms